DL/News Research ChatBot using AWS BedRock/requirements1.txt
numpy
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.51
botocore==1.35.51
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
compressed-tensors==0.6.0
cryptography==43.0.3
dataclasses-json==0.5.14
datasets==3.0.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
emoji==2.14.0
entrypoints==0.4
eval_type_backport==0.2.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.4
langchain-aws==0.2.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.137
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
mistral_common==1.4.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numexpr==2.10.1
nvidia-ml-py==12.560.30
olefile==0.47
openai==1.53.0
opencv-python-headless==4.10.0.84
orjson==3.10.10
outlines==0.0.46
packaging==24.1
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
pillow==10.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pypdf==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
ray==2.38.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
s3transfer==0.10.3
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
streamlit==1.22.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.1
toml==0.10.2
toolz==1.0.0
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.6
transformers==4.46.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
unstructured==0.9.2
unstructured-client==0.26.2
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
validators==0.34.0
#vllm==0.6.3.post1
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.0
zipp==3.20.2
